+++ @vue-start/element-pro/src/comp/Select.tsx
@@ -1,11 +1,19 @@
-import { defineComponent, ExtractPropTypes, PropType, ref } from "vue";
+import { computed, defineComponent, ExtractPropTypes, PropType, ref } from "vue";
 import { ElSelect, ISelectProps, ElOption, IOptionProps } from "element-plus";
 import { TOption } from "@vue-start/pro";
-import { keys, map, omit } from "lodash";
+import { get, isArray, isString, keys, map, omit } from "lodash";
 import { createExposeObj } from "@vue-start/pro";
+import { formatValue, parseValue } from "@vue-start/hooks";
 
 const proSelectProps = () => ({
   options: Array as PropType<Array<TOption & IOptionProps>>,
+  fieldNames: { type: Object },
+  // **************** expose 拓展 **********************
+  expMethods: { type: Array as PropType<string[]>, default: () => ["focus", "blur", "selectedLabel"] },
+  // **************** value 格式自定义 **********************
+  separator$: { type: String }, //分割符
+  parseValue$: { type: Function }, //解析value 优先级最高
+  formatValue$: { type: Function }, //转换value 优先级最高
 });
 
 export type ProSelectProps = Partial<ExtractPropTypes<ReturnType<typeof proSelectProps>>> & ISelectProps;
@@ -18,21 +26,52 @@ export const ProSelect = defineComponent<ProSelectProps>({
   setup: (props, { slots, emit, expose }) => {
     const originRef = ref();
 
-    expose(createExposeObj(originRef));
+    expose(createExposeObj(originRef, props.expMethods));
 
-    const invalidKeys = keys(proSelectProps());
+    const reOptions = computed(() => {
+      return map(props.options, (item) => {
+        return {
+          ...item,
+          label: get(item, props.fieldNames?.label || "label"),
+          value: get(item, props.fieldNames?.value || "value"),
+          disabled: get(item, props.fieldNames?.label || "disabled"),
+        };
+      });
+    });
+
+    const modelValue = computed(() => {
+      const mv = props.modelValue;
+
+      if (props.parseValue$) return props.parseValue$(mv, props);
+
+      return parseValue(mv, { multiple: props.multiple, separator: props.separator$ });
+    });
+
+    const handleChange = (v: any) => {
+      if (props.formatValue$) {
+        emit("update:modelValue", props.formatValue$(v, props));
+        return;
+      }
+
+      if (props.separator$) {
+        emit("update:modelValue", formatValue(v, { multiple: props.multiple, separator: props.separator$ }));
+        return;
+      }
+      emit("update:modelValue", v);
+    };
+
+    const invalidKeys = [...keys(proSelectProps()), "modelValue"];
     return () => {
       return (
         <ElSelect
           ref={originRef}
           {...omit(props, invalidKeys)}
-          onUpdate:modelValue={(v) => {
-            emit("update:modelValue", v ? v : undefined);
-          }}
+          modelValue={modelValue.value}
+          onUpdate:modelValue={(v) => handleChange(v)}
           v-slots={omit(slots, "default")}>
           {slots.start?.()}
 
-          {map(props.options, (item: TOption & IOptionProps) => {
+          {map(reOptions.value, (item: TOption & IOptionProps) => {
             //插槽重写label
             const labelEl = slots.label?.(item);