+++ @vue-start/hooks/src/util/options.ts
@@ -1,5 +1,6 @@
-import { forEach, get, map, omit, reduce, size } from "lodash";
+import { forEach, get, isArray, isNumber, isString, join, last, map, omit, reduce, size } from "lodash";
 import { FieldNames, TOptions, TreeOptions } from "@vue-start/pro";
+import { findTreeItem, findTreeItem2 } from "./collection";
 
 export const getFieldNames = (fieldNames?: FieldNames) => {
   const labelName = fieldNames?.label || "label";
@@ -90,4 +91,148 @@ export const treeToOptionsMap = (
       treeToOptionsMap(children, fieldNames, mapObj, itemMapObj);
     }
   });
+};
+
+/* ************************** value ******************************* */
+
+export type TValueOpts = {
+  multiple?: boolean;
+  allPath?: boolean; //tree数据使用完整路径
+  separator?: string; //分割符
+  itemSeparator?: string; //子分割符
+};
+
+/**
+ * 解析
+ */
+export const parseValue = (v: any, opts: TValueOpts) => {
+  //
+  if (!v) {
+    return opts.multiple ? [] : v;
+  }
+
+  //已经是数组格式
+  if (isArray(v)) return v;
+
+  const separator = opts.separator || ",";
+  const itemSeparator = opts.itemSeparator || "/";
+
+  //字符串情况
+  if (isString(v)) {
+    if (opts.multiple) {
+      if (opts.allPath) {
+        return map(v.split(separator), (item) => {
+          if (item) return item.split(itemSeparator);
+          return [];
+        });
+      }
+      return v.split(separator);
+    } else {
+      if (opts.allPath) {
+        return v.split(itemSeparator);
+      }
+    }
+  }
+
+  return v;
+};
+
+/**
+ * 转换
+ */
+export const formatValue = (v: any, opts: TValueOpts) => {
+  if (!isArray(v)) return v;
+
+  const separator = opts.separator || ",";
+  const itemSeparator = opts.itemSeparator || "/";
+
+  if (opts.multiple) {
+    if (opts.allPath) {
+      return join(
+        map(v, (item) => {
+          return item ? join(item, itemSeparator) : "";
+        }),
+        separator,
+      );
+    } else {
+      return join(v, separator);
+    }
+  } else {
+    if (opts.allPath) {
+      return join(v, itemSeparator);
+    }
+  }
+
+  return v;
+};
+
+export type TLabelOpts = TValueOpts & {
+  options?: Record<string, any>[];
+  fieldNames?: Record<string, string>;
+  showAllPath?: boolean;
+};
+
+/**
+ * 获取value对应的record
+ */
+export const findValueRecord = (v: any, opts: TLabelOpts) => {
+  const rv = parseValue(v, opts);
+
+  if (!rv) return null;
+
+  const valueName = opts.fieldNames?.value || "value";
+  const childrenName = opts.fieldNames?.children || "children";
+
+  if (opts.multiple) {
+    if (opts.allPath) {
+      return map(rv, (item) => {
+        return findTreeItem2(opts.options || [], item, { value: valueName, children: childrenName }).parentList;
+      });
+    } else {
+      return map(rv, (item) => {
+        return findTreeItem(
+          opts.options || [],
+          (subItem) => get(subItem, valueName) === item,
+          { children: childrenName },
+          [],
+        ).parentList;
+      });
+    }
+  } else {
+    if (opts.allPath) {
+      return findTreeItem2(opts.options || [], rv, { value: valueName, children: childrenName }).parentList;
+    }
+  }
+
+  return findTreeItem(opts.options || [], (item) => get(item, valueName) === rv, { children: childrenName }, [])
+    .parentList;
+};
+
+/**
+ * 获取value对应的label
+ */
+export const findValueLabel = (r: any, opts: TLabelOpts) => {
+  if (!r) return "";
+  if (isArray(r) && size(r) <= 0) return "";
+
+  const labelName = opts.fieldNames?.label || "label";
+
+  const separator = opts.separator || ",";
+  const itemSeparator = opts.itemSeparator || "/";
+
+  return join(
+    map(r, (sr) => {
+      if (isArray(sr)) {
+        if (opts.showAllPath) {
+          return join(
+            map(sr, (item) => get(item, labelName)),
+            itemSeparator,
+          );
+        }
+        return get(last(sr), labelName);
+      }
+      return get(sr, labelName);
+    }),
+    separator,
+  );
 };