+++ src/platform-element-plus/index.ts
@@ -8,6 +8,7 @@ import { createRequest, ContentTypeInterceptor } from "@vue-start/request";
 
 import '@vue-start/css/normalize.css'
 import "@vue-start/pro/index.css";
+import "@vue-start/element-pro/index.css";
 import "@vue-start/media/index.css";
 
 import "@/style/index.css";