+++ src/views/el/cascader/index.tsx
@@ -1,17 +1,11 @@
 import { defineComponent } from "vue";
 
-import Basic from "./demo/basic";
-import Slot from "./demo/slot";
-import Value from "./demo/value";
 import Compose from "./demo/compose";
 
 export default defineComponent(() => {
   return () => {
     return (
       <>
-        <Basic />
-        <Slot />
-        <Value />
         <Compose />
       </>
     );